Move server split into a real server/ subpackage (D172)
The flat fused_render/_server_*.py layout from the previous commit was
a naming convention standing in for a package structure: 16 files
stuttering the same _server_ prefix, sitting loose next to unrelated
modules. Moved into fused_render/server/ with a routers/ subfolder for
the route-wiring-only modules (shell, config, fs_read, render, run,
export); files that carry real reusable logic alongside their route(s)
(fs_mutate, session, ai) stay at the server/ top level. server/app.py
is create_app() itself; server/__init__.py re-exports create_app /
set_server_origin_env / export_app_env so every existing
`from fused_render.server import create_app` call site needed no change.

Mechanical rename + import-path rewrite, same byte-identical-content
discipline as the split itself. Fixed one real bug the move surfaced:
STATIC_DIR was computed from common.py's own __file__, which is now one
directory deeper — repointed at the fused_render package root instead.
Full suite still matches baseline: 2691 passed, 86 skipped, 1
pre-existing unrelated failure.
