+++ DECISIONS.md
@@ -538,6 +538,12 @@ Security layer (token, Origin/Host validation, sandboxed bridge — see threat n
 
 *(Bookkeeping: THIRD same-day collision on this range. `main` had already renumbered the fused.ai rows to D168/D169 above — itself a fix for colliding with D166 — while this branch independently took D167/D168/D169. D167 is untouched on `main`, so the binder row keeps it and its `_binding.py` citation stays valid; the other two renumber to **D170/D171**, with the CI workflow comment updated to match. The root cause is structural and unfixed: two branches cut from one base always pick the same "next" number, and nothing here makes that detectable before a merge.)*
 
+### server.py split into a real package (2026-07-29)
+
+| # | Decision | Choice | Rationale / rejected alternatives |
+|---|---|---|---|
+| D172 | **`server.py` (4900 lines, every concern + a 1180-line `create_app()` with every route inline) is split into `fused_render/server/` — one module per concern, `routers/` for route-wiring-only modules, routes as real `APIRouter`s instead of closures** | `server.py` had grown to mix git-ignore-aware walking, directory listing, a stat cache, an fs watch registry, remote-mount probing, the fused.ai Claude session, template/condition resolution, raw-proxy hardening, fs mutation endpoints, and route wiring in one file. Split into `server/common.py`, `templates.py` (+ condition gating), `gitignore.py`, `walk.py`, `mount.py` (+ stat cache), `watch.py`, `proxy.py`, `fs_mutate.py`, `session.py`, `ai.py`, plus `server/routers/{shell,config,fs_read,render,run,export}.py` for the route groups that carry no reusable logic of their own — `server/app.py` is left with just `create_app()`: `FastAPI()` construction, `app.state` wiring, and `include_router()` calls. `server/__init__.py` re-exports `create_app`/`set_server_origin_env`/`export_app_env` so every existing `from fused_render.server import create_app` call site (cli.py, app.py, ~30 tests) needed no change. First cut used a flat `fused_render/_server_*.py` naming scheme (no subpackage) — correctly called out as not how this should be organized (a stuttering `_server_` prefix on 16 files sitting loose next to unrelated modules is a naming convention standing in for a package structure) and redone as the `server/`+`routers/` layout above. Every moved line is byte-identical to the original (a script diffed the extraction against `git show HEAD:fused_render/server.py`, not eyeballed); the only actual code changes are the closure→`Depends()`/`app.state` conversions `APIRouter` requires (documented per-route) and one real bug the split surfaced: `_STAT_CACHE_GEN` was a plain module-level int incremented via `global` in `mount.py` — imported by value into whichever module read it, that copy would freeze at 0 forever (a rebind via `global` only updates the defining module's own `__dict__`), silently breaking the `/api/fs/stat` TOCTOU race guard (caught by `test_midflight_invalidation_is_not_refilled`, not by inspection). Fixed by reading it live through `mount._STAT_CACHE_GEN` (module-qualified) at the two call sites in `fs_read.py` instead of an imported bare name. Rejected: re-export shims in `server/__init__.py` to keep old `server.X`/`monkeypatch.setattr(server, "X", ...)` call sites working unchanged — tests were retargeted to wherever their code actually lives instead, since a shim would have quietly reintroduced the exact `_STAT_CACHE_GEN`-style bug for every moved mutable global. Full suite matches the pre-split baseline exactly: 2691 passed, 86 skipped, 1 pre-existing unrelated failure (`test_an_abandoned_merge_closes_the_day_s_files` needs `/proc/self/fd`, Linux-only). |
+
 ## Open items (small, non-blocking)
 
 - Shell visual design: v1 = clean minimal; real design pass later.
