+++ fused_render/server/routers/render.py
@@ -4,8 +4,8 @@ import urllib.request
 from fastapi import APIRouter
 from fastapi.responses import HTMLResponse
 
-from fused_render._server_common import _error
-from fused_render._server_session import _is_file_mount_safe
+from fused_render.server.common import _error
+from fused_render.server.session import _is_file_mount_safe
 from fused_render.shell import mounts as shell_mounts
 
 router = APIRouter()
