+++ fused_render/server/session.py
@@ -4,7 +4,7 @@ from urllib.parse import parse_qsl, urlsplit
 from fastapi import APIRouter, Body, Header
 from fused_render.shell import storage
 
-from fused_render._server_common import _error, _require_fused
+from fused_render.server.common import _error, _require_fused
 
 router = APIRouter()
