+++ tests/test_canvas_template.py
@@ -18,7 +18,7 @@ import sys
 
 import pytest
 
-from fused_render import _server_templates
+from fused_render.server import templates as _server_templates
 
 
 TEMPLATES_DIR = _server_templates.TEMPLATES_DIR
