+++ tests/test_condition_gate_seed.py
@@ -22,7 +22,7 @@ from fastapi.testclient import TestClient
 
 import fused_render.shell.mounts as mounts_mod
 from fused_render import server
-from fused_render import _server_templates
+from fused_render.server import templates as _server_templates
 
 MOUNT_PREFIX = "/fake-mounts/"
 STORE = "/fake-mounts/s3demo/store"
