+++ tests/test_condition_probe_budget.py
@@ -22,7 +22,7 @@ from fastapi.testclient import TestClient
 
 import fused_render.server as server
 import fused_render.shell.mounts as mounts_mod
-from fused_render import _server_templates
+from fused_render.server import templates as _server_templates
 from fused_render.server import create_app
 from test_shell_mounts import StubRcd
