+++ tests/test_fs_walk_bound.py
@@ -6,8 +6,9 @@ import asyncio
 
 from fastapi.testclient import TestClient
 
-from fused_render import _server_fs_read, server
-from fused_render._server_fs_read import _walk_bfs
+from fused_render.server.routers import fs_read as _server_fs_read
+from fused_render import server
+from fused_render.server.routers.fs_read import _walk_bfs
 from fused_render.server import create_app
